app/requirements.txt
flask>=2.0.1,<3.0.0
marshmallow==3.22.0
requests>=2.7.0
websocket-client>=1.0.0
tqdm>=4.62.3,<5.0.0
werkzeug>=2.0.3
gevent>=22.10.2,<23.0.0
pyzmq>=25.1.1,<26.0.0
openfabric-pysdk==0.2.9
chromadb>=1.0.6
streamlit>=1.32.0
Pillow>=10.2.0